Move design time initializer methods to Helper class
