+++ src/Games/NexusMods.Games.AdvancedInstaller.UI/Content/Right/SelectLocation/SelectLocationDesignViewModel.cs
@@ -8,5 +8,5 @@ namespace NexusMods.Games.AdvancedInstaller.UI.SelectLocation;
 [ExcludeFromCodeCoverage]
 internal class SelectLocationDesignViewModel : SelectLocationViewModel
 {
-    public SelectLocationDesignViewModel() : base(BodyDesignViewModel.CreateDesignGameLocationsRegister(), null) { }
+    public SelectLocationDesignViewModel() : base(DesignTimeHelpers.CreateDesignGameLocationsRegister(), null) { }
 }
