+++ src/Games/NexusMods.Games.AdvancedInstaller.UI/Pages/AdvancedInstallerPages/AdvancedInstallerPageDesignViewModel.cs
@@ -1,16 +1,17 @@
 ﻿using System.Diagnostics.CodeAnalysis;
+using NexusMods.DataModel.Games;
+using NexusMods.DataModel.ModInstallers;
+using NexusMods.Paths;
+using NexusMods.Paths.FileTree;
 
 namespace NexusMods.Games.AdvancedInstaller.UI;
 
 [ExcludeFromCodeCoverage]
 // ReSharper disable once UnusedType.Global
-public class AdvancedInstallerPageDesignViewModel : AViewModel<IAdvancedInstallerPageViewModel>,
-    IAdvancedInstallerPageViewModel
+internal class AdvancedInstallerPageDesignViewModel : AdvancedInstallerPageViewModel
 {
-    public IFooterViewModel FooterViewModel { get; } = new FooterDesignViewModel();
-    public IBodyViewModel BodyViewModel { get; } = new BodyDesignViewModel();
-
-    public bool ShouldInstall { get; } = false;
-
-    public string ModName { get; set; } = "Design Mod Name";
+    public AdvancedInstallerPageDesignViewModel() : base(
+        "Design Mod Name",
+        DesignTimeHelpers.CreateDesignFileTree(),
+        DesignTimeHelpers.CreateDesignGameLocationsRegister()) { }
 }
